# service/src/main/java/at/uni/innsbruck/htibot/rest/generated/model/LanguageEnum.java
package at.uni.innsbruck.htibot.rest.generated.model;

import io.swagger.annotations.ApiModel;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonValue;

/**
 * Enumeration of supported languages
 */
public enum LanguageEnum {
  
  ENGLISH("English"),
  
  FRENCH("French"),
  
  ITALIAN("Italian"),
  
  GERMAN("German");

  private String value;

  LanguageEnum(String value) {
    this.value = value;
  }

    /**
     * Convert a String into String, as specified in the
     * <a href="https://download.oracle.com/otndocs/jcp/jaxrs-2_0-fr-eval-spec/index.html">See JAX RS 2.0 Specification, section 3.2, p. 12</a>
     */
	public static LanguageEnum fromString(String s) {
      for (LanguageEnum b : LanguageEnum.values()) {
        // using Objects.toString() to be safe if value type non-object type
        // because types like 'int' etc. will be auto-boxed
        if (java.util.Objects.toString(b.value).equals(s)) {
          return b;
        }
      }
      throw new IllegalArgumentException("Unexpected string value '" + s + "'");
	}
	
  @Override
  @JsonValue
  public String toString() {
    return String.valueOf(value);
  }

  @JsonCreator
  public static LanguageEnum fromValue(String value) {
    for (LanguageEnum b : LanguageEnum.values()) {
      if (b.value.equals(value)) {
        return b;
      }
    }
    throw new IllegalArgumentException("Unexpected value '" + value + "'");
  }
}