# teams/apiclient/src/model/BaseErrorModel.js
/**
 * HTI Bot API
 * Interact with the HTI Bot Backend
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';

/**
 * The BaseErrorModel model module.
 * @module model/BaseErrorModel
 * @version 1.0.0
 */
class BaseErrorModel {
    /**
     * Constructs a new <code>BaseErrorModel</code>.
     * @alias module:model/BaseErrorModel
     */
    constructor() { 
        
        BaseErrorModel.initialize(this);
    }

    /**
     * Initializes the fields of this object.
     * This method is used by the constructors of any subclasses, in order to implement multiple inheritance (mix-ins).
     * Only for internal use.
     */
    static initialize(obj) { 
    }

    /**
     * Constructs a <code>BaseErrorModel</code> from a plain JavaScript object, optionally creating a new instance.
     * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @param {module:model/BaseErrorModel} obj Optional instance to populate.
     * @return {module:model/BaseErrorModel} The populated <code>BaseErrorModel</code> instance.
     */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new BaseErrorModel();

            if (data.hasOwnProperty('resultCode')) {
                obj['resultCode'] = ApiClient.convertToType(data['resultCode'], 'Number');
            }
            if (data.hasOwnProperty('message')) {
                obj['message'] = ApiClient.convertToType(data['message'], 'String');
            }
        }
        return obj;
    }

    /**
     * Validates the JSON data with respect to <code>BaseErrorModel</code>.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @return {boolean} to indicate whether the JSON data is valid with respect to <code>BaseErrorModel</code>.
     */
    static validateJSON(data) {
        // ensure the json data is a string
        if (data['message'] && !(typeof data['message'] === 'string' || data['message'] instanceof String)) {
            throw new Error("Expected the field `message` to be a primitive type in the JSON string but got " + data['message']);
        }

        return true;
    }


}



/**
 * HTTP Status Code
 * @member {Number} resultCode
 */
BaseErrorModel.prototype['resultCode'] = undefined;

/**
 * Error Message
 * @member {String} message
 */
BaseErrorModel.prototype['message'] = undefined;






export default BaseErrorModel;