# teams/apiclient/src/model/LanguageEnum.js
/**
 * HTI Bot API
 * Interact with the HTI Bot Backend
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';
/**
* Enum class LanguageEnum.
* @enum {}
* @readonly
*/
export default class LanguageEnum {
    
        /**
         * value: "English"
         * @const
         */
        "English" = "English";

    
        /**
         * value: "French"
         * @const
         */
        "French" = "French";

    
        /**
         * value: "Italian"
         * @const
         */
        "Italian" = "Italian";

    
        /**
         * value: "German"
         * @const
         */
        "German" = "German";

    

    /**
    * Returns a <code>LanguageEnum</code> enum value from a Javascript object name.
    * @param {Object} data The plain JavaScript object containing the name of the enum value.
    * @return {module:model/LanguageEnum} The enum <code>LanguageEnum</code> value.
    */
    static constructFromObject(object) {
        return object;
    }
}